requirements.txt
amqp==5.3.1
annotated-types==0.7.0
anyio==4.9.0
appdirs==1.4.4
billiard==4.2.1
blinker==1.9.0
cachetools==5.5.2
celery==5.5.2
certifi==2025.4.26
charset-normalizer==3.4.2
click==8.2.0
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
colormath==3.0.0
contourpy==1.3.2
cycler==0.12.1
dateparser==1.2.1
Deprecated==1.2.18
Flask==3.1.1
flask-cors==6.0.0
Flask-Limiter==3.12
fonttools==4.58.0
google-auth==2.40.1
google-genai==1.15.0
greenlet==3.2.2
gunicorn==23.0.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
importlib_metadata==8.7.0
itsdangerous==2.2.0
Jinja2==3.1.6
joblib==1.5.0
kiwisolver==1.4.8
kombu==5.5.3
limits==5.2.0
markdown-it-py==3.0.0
MarkupSafe==3.0.2
matplotlib==3.10.3
mdurl==0.1.2
networkx==3.4.2
numpy==2.2.5
ordered-set==4.1.0
packaging==25.0
parsedatetime==2.6
pdf2image==1.17.0
pgvector==0.4.1
pillow==11.2.1
playwright==1.52.0
prompt_toolkit==3.0.51
psycopg2-binary==2.9.10
pyasn1==0.6.1
pyasn1_modules==0.4.2
pydantic==2.11.4
pydantic_core==2.33.2
pyee==11.1.1
Pygments==2.19.1
PyJWT==2.10.1
PyMuPDF==1.26.0
pyparsing==3.2.3
pyppeteer==2.0.0
pyppeteer_stealth==2.7.4
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
pytz==2025.2
PyVirtualDisplay==3.0
redis==6.1.0
regex==2024.11.6
requests==2.32.3
rich==13.9.4
rsa==4.9.1
scikit-learn==1.6.1
scipy==1.15.3
six==1.17.0
smtp2go==2.3.1
sniffio==1.3.1
SQLAlchemy==2.0.41
threadpoolctl==3.6.0
tqdm==4.67.1
typing-inspection==0.4.0
typing_extensions==4.13.2
tzdata==2025.2
tzlocal==5.3.1
urllib3==1.26.20
vine==5.1.0
watchdog==6.0.0
wcwidth==0.2.13
websockets==10.4
Werkzeug==3.1.3
wrapt==1.17.2
zipp==3.22.0
